Feedback 2
[main.cpp]
— Metadata Class moved to Generator.cpp/.h.

[Generator.cpp/.h]
— DumpEditorOnlyMetadata() -> DumpEditorOnly() and is part of new Metadata class now.
